requirements.txt
# ── Core ML ───────────────────────────────────────────────────
xgboost>=3.2.0
scikit-learn>=1.8.0
pandas>=2.3.0
numpy>=2.4.0

# ── Experiment tracking ───────────────────────────────────────
mlflow>=3.11.0
dagshub>=0.7.0

# ── Data validation ───────────────────────────────────────────
great-expectations>=1.17.0

# ── Config ────────────────────────────────────────────────────
pyyaml>=6.0

# ── Security scanning (Phase 3) ───────────────────────────────
bandit>=1.9.0
safety>=3.7.0
